Processamento de atividades. alteracao de tempo.

+++ source/src/main/java/br/com/sysmap/taskqueue/model/tipos/OperacaoTempo.java
@@ -3,18 +3,18 @@ package br.com.sysmap.taskqueue.model.tipos;
 public enum OperacaoTempo {
 	SOMA() {
 		@Override
-		public Integer calcular(Integer tempo) {
+		public Long calcular(Long tempo) {
 			return tempo += 10;
 		}
 	},
 	SUBTRACAO {
 		@Override
-		public Integer calcular(Integer tempo) {
+		public Long calcular(Long tempo) {
 			return tempo -= 10;
 		}
 	};
 	
-	public abstract Integer calcular(Integer tempo);
+	public abstract Long calcular(Long tempo);
